Maintainability... Massively refactored MainWindow so most widgets and actions get setup by BEAGLESMainWindow parent class
