+++ libs/ui/callbacks/boundingBoxCallbacks.py
@@ -1,0 +1,48 @@
+from libs.shape import Shape, DEFAULT_LINE_COLOR, DEFAULT_FILL_COLOR
+
+
+class BoundingBoxCallbacks:
+    def hideAll(self):
+        self.togglePolygons(False)
+
+    def showAll(self):
+        self.togglePolygons(True)
+
+    def delBox(self):
+        self.remLabel(self.canvas.deleteSelected())
+        self.setDirty()
+        if self.noShapes():
+            for action in self.actions.onShapesPresent:
+                action.setEnabled(False)
+        # noinspection PyTypeChecker
+
+    def shapeFillColor(self):
+        color = self.colorDialog.getColor(self.fillColor, u'Choose fill color',
+                                          default=DEFAULT_FILL_COLOR)
+        if color:
+            self.canvas.selectedShape.fill_color = color
+            self.canvas.update()
+            self.setDirty()
+
+    def shapeLineColor(self):
+        color = self.colorDialog.getColor(self.lineColor, u'Choose line color',
+                                          default=DEFAULT_LINE_COLOR)
+        if color:
+            self.canvas.selectedShape.line_color = color
+            self.canvas.update()
+            self.setDirty()
+
+    def boxLineColor(self):
+        color = self.colorDialog.getColor(self.lineColor, u'Choose line color',
+                                          default=DEFAULT_LINE_COLOR)
+        if color:
+            self.lineColor = color
+            Shape.line_color = color
+            self.canvas.setDrawingColor(color)
+            self.canvas.update()
+            self.setDirty()
+
+    def copySelectedShape(self):
+        self.addLabel(self.canvas.copySelectedShape())
+        # fix copy and delete
+        self.shapeSelectionChanged(True)
